commit pos deploy

+++ app/Http/Controllers/JogoController.php
@@ -1,17 +1,22 @@
 <?php namespace flanacao\Http\Controllers;
 
 use Illuminate\Support\Facades\DB;
+//use Illuminate\Support\Facades\Auth;
 use Request;
 use Redirect;
 
 class JogoController extends Controller {
 
   function __construct(){
-    # code...
+    $this->middleware('auth', 
+      ['except' => ['classificacao']]);
   }
 
 // ******************************************************************
   public function lista() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $jogos = DB::select("Select j.*, t1.nome as mandante, t2.nome as visitante FROM jogos j, times t1, times t2
       WHERE (j.time1_id = t1.id) AND (j.time2_id = t2.id) ORDER BY data_jogo desc");
 
@@ -20,6 +25,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function salvar_resultados() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $jogo = $_GET['j'];
     $pMandante = $_GET['pm'];
     $pVisitante = $_GET['pv'];
@@ -36,6 +44,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function cadastrar() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $times = DB::select("Select * FROM times WHERE ativo = 1");
     $campeonatos = DB::select("Select * FROM campeonatos WHERE ativo = 1 ORDER BY nome");
 
@@ -49,6 +60,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function novo() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $mandante = Request::input('mandante');
     $visitante = Request::input('visitante');
     $campeonato = Request::input('campeonato');
@@ -58,10 +72,10 @@ class JogoController extends Controller {
     $hora_jogo = $hora_jogo1.":".$hora_jogo2;
 
     if ($campeonato == "0") 
-      return Redirect::back()->with('err', 'ERRO: Informar o campeonato.'); 
+      return Redirect::back()->with('err', 'ERRO: Informar o campeonato.')->withInput(Request::all()); 
 
     if ($mandante === $visitante)
-      return Redirect::back()->with('err', 'ERRO: Informar times diferentes.');
+      return Redirect::back()->with('err', 'ERRO: Informar times diferentes.')->withInput(Request::all());
 
     $id = DB::table('jogos')->insertGetId(
       [
@@ -77,6 +91,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   private function cria_registros($id) {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $users = DB::select("SELECT id FROM users WHERE inscricao_liberada = 1");
     foreach ($users as $u) {
       DB::table('palpites')->insert(
@@ -87,6 +104,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function editar() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $mandante = Request::input('mandante');
     $visitante = Request::input('visitante');
     $campeonato = Request::input('campeonato');
@@ -118,6 +138,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function mudarLiberado() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $jogo = $_GET['jogo'];
     $flag = $_GET['flag'];
 
@@ -129,6 +152,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function excluir() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $jogo = $_GET['jogo'];
     DB::table('jogos')->where('id', '=', $jogo)->delete();
     DB::table('palpites')->where('jogo_id', '=', $jogo)->delete();
@@ -136,7 +162,7 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function listarPalpites() {
-    $user = 1; //pegar esse valor da sessão
+    $user = \Auth::user()->id; 
 
     //verifica se os palpites não estão bloqueados
     $config = DB::table('configuracoes')->first();
@@ -150,7 +176,7 @@ class JogoController extends Controller {
       return ("Você está impedido de realizar palpites.<br />
         Entre em contato com o administrador do bolão. <br /><a href='/'>voltar</a>");
     
-    $jogos = DB::select("SELECT j.*, t1.nome as mandante, t2.nome as visitante, p.palpite_mandante, p.palpite_visitante  
+    $jogos = DB::select("SELECT j.*, t1.nome as mandante, t1.arquivo as escudo1, t2.nome as visitante, t2.arquivo as escudo2, p.palpite_mandante, p.palpite_visitante  
       FROM jogos j, times t1, times t2, palpites p  
       WHERE (t1.id = j.time1_id) AND (t2.id = j.time2_id) AND (j.liberado = 1) AND (p.jogo_id = j.id) AND (p.usuario_id = $user)");
 
@@ -182,6 +208,9 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function calcularPontuacoes() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
     $config = DB::table('configuracoes')->first();
     $placarCompleto = $config->pontuacao_placar_completo;
     $resultado = $config->pontuacao_resultado;
@@ -236,9 +265,60 @@ class JogoController extends Controller {
 
 // ******************************************************************
   public function classificacao() {
-    $clas = DB::select("SELECT SUM(p.pontuacao) as pontuacao, p.usuario_id, u.name FROM palpites p, users u WHERE (u.id = p.usuario_id) GROUP BY usuario_id");
+    $pontuacoes = DB::select("SELECT * FROM pontuacoes"); 
+    $clas = DB::select("SELECT SUM(p.pontuacao) as pontuacao, p.usuario_id, u.name, u.login FROM palpites p, users u WHERE (u.id = p.usuario_id) GROUP BY usuario_id ORDER BY pontuacao DESC");
 
-    return view('classificacao', ['clas' => $clas]);
+    $ids = array();
+    $extras = array();
+    // soma a pontuação extra a pontuação normal
+    //if (!empty($pontuacoes)) { // verifica se a tabela 'pontuacoes' está vazia
+      foreach ($clas as $reg) {
+        $id = $reg->usuario_id;
+        $pont = DB::select("SELECT sum(pontuacao) as p FROM pontuacoes 
+          WHERE usuario_id = $id");
+          foreach ($pont as $value) { 
+            if ($value->p != 'NULL')
+              $reg->pontuacao += ($value->p); 
+          }   
+
+        //carrega os detalhamentos dos palpites
+        $ids[$id] = DB::select("SELECT t1.nome as mandante, t2.nome as visitante, j.placar1, j.placar2, p.palpite_mandante, p.palpite_visitante, p.pontuacao 
+          FROM times t1, times t2, palpites p, jogos j 
+          WHERE (t1.id = j.time1_id) AND (t2.id = j.time2_id) AND (p.jogo_id = j.id) AND 
+          (p.usuario_id = $id) AND (j.placar1 >= 0) AND (p.palpite_mandante >= 0) 
+          ORDER BY j.data_jogo DESC");
+
+        $extras[$id] = DB::select("SELECT * FROM pontuacoes WHERE usuario_id = $id");
+      }
+
+      usort($clas, function($a, $b) {
+        return strcmp($b->pontuacao, $a->pontuacao);
+      });
+    //}
+
+    return view('classificacao', ['clas' => $clas, 'ids' => $ids, 'extras' => $extras]);
+  }
+
+  // ******************************************************************
+  public function listaOpcoes() {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
+    $config = DB::table('configuracoes')->first();
+    return view('jogos.bloqueio')->withConfig($config);
+  }
+
+  // ******************************************************************
+  public function bloquearGeral($valor) {
+    if (\Auth::user()->adm != 1) // não é ADM
+      return redirect('/classificacao');
+
+    if ($valor == 1) $valor = 0;
+    else $valor = 1;
+    DB::table('configuracoes')
+      ->update(['bloquear_palpites' => $valor]);
+
+    return redirect('/jogos/resultados');
   }
